requirements.txt
python-dotenv
requests
openai
pytz
pytest
pytest-asyncio
pyfakefs
tzdata